Set sequence value to skip hardcoded ids
When running up the collection exercise service locally
BRES_collectionexercise_seeddata.sql inserts a collection exercise with
a hardcoded id. This means the sequence doesn't get incremented
therefore the next time the api tries to insert a collection exercise
using the sequence it will fail.

This change should have any impact in production as it will either set
the sequence to what it currently is because it's greater than the 4.

The reason it has been set to 4 is because there are a couple more rows
in the dev database