Changed error handling to avoid infinite retry
Originally the message would be left on the queue if an error occurred
handling an incoming message on the Seft.Instruments queue.  This would
cause the message to be processed again and the error most likely
repeated.

This PR covers a change to remove the message from the Seft.Instruments
queue and drop an error to the Seft.Instruments.DLQ queue.