Add wait in pre-integration-test (#76)
* Add wait in pre-integration-test

The integration tests are hit and miss and it's likely a timing issue
with action taking too long to start. We were waiting an arbitrary 10s
which might not have been enough.

This change uses a different maven plugin to check the healthcheck/info
api waiting up to 60 seconds for it to return a 200.