Create Action Rules when Events Created
Currently there is nothing to automatically create the action rules that
cause letters and emails to be sent when a collection exercise's events
are created. This will fix that so letters & emails will be sent without
manual intervention.

Updating these dates is out of scope for this ticket.

https://trello.com/c/Eknzy8H2/119-tbl123-link-action-rules-to-action-plans-at-create-time-create
Dependant on: ONSdigital/rm-action-service#69