Fix social survey event dates
Currently when you create a social survey's collection exercise event
dates there is an exception. This is caused by assuming all
surveys are business surveys when creating action rules. This
adds a check for if the survey is a business survey, removing the
automatic creation of action rules for non-business survey types.