Add authentication header & remove unnecessary survey calls (#112)
Call to action service was being done without an authentication header
resulting in 401s on event update. This pr adds the header into the
request. Also a number of unnecessary calls are being made to survey
service, changed the logic order so the calls are not being made
unless it is nedded.