Fix event validation when adding and updating events (#115)
* Fix event validation when adding and updating events

Currently the validation is not flexible enough to allow validation
on  both update and creation. This PR makes the validation work with
both create events and update events. Also making sure all business
rules are captured within the validation.

* Remove 24hrs apart check and furture check

To allow ATs to provision events the checks for dates being 24hrs
apart and the dates are in the future will be moved into the UI.
(response-operations)

* Refactor event validators into types

Before we had one validator 'businessEventvalidator' but the
validators can be more fine grained. So refactored the validator
MandatoryEventValidator, ReminderEventValidator and
ReferncePeriodEventValidator.

* Fix flakey integration test

The integration test to update events currently fails when comparing
two timeStamps, which are off by less than a 100th of a second. This
change rounds the timestamps to the nearest minute before comparing.