Adjust reminder event validation
Currently reminder events that are in the past can not be updated,
which is a valid validation, but this check should only apply if the
collection exercise has been executed. This PR adds a check for CE
state before validating the reminder event has passed. Allowing a user
to update reminder events that have passed but not actually created any
notification files/emails.